fix: review, mlReviews, list RESTful verbs + frontend correctly uses HATEOAS

+++ webapp/src/main/java/ar/edu/itba/paw/webapp/controller/ReviewController.java
@@ -239,21 +239,32 @@ public class ReviewController {
         boolean liked = reviewService.likeReview(id, ReviewTypes.REVIEW_MEDIA);
         if (liked)
             return Response.ok("{\"message\":\"Successfully liked list.\"}").build();
-        return Response.status(Response.Status.BAD_REQUEST)
+        return Response.status(Response.Status.CONFLICT)
                 .entity("{\"message\":\"List is already liked.\"}").build();
     }
 
     @DELETE
-    @Path("/{id}/likes")
+    @Path("/{id}/likes/{username}")
     @PreAuthorize("@accessValidator.isUserLoggedIn()")
     @Produces(MediaType.APPLICATION_JSON)
-    public Response deleteReviewLike(@PathParam("id") int id) {
-
-        boolean removed = reviewService.removeLikeReview(id, ReviewTypes.REVIEW_MEDIA);
-        if (removed)
-            return Response.noContent().build();
-        return Response.status(Response.Status.BAD_REQUEST)
-                .entity("{\"message\":\"List is not liked.\"}").build();
+    public Response deleteReviewLike(@PathParam("id") int id,
+                                     @PathParam("username") String username) {
+        try{
+            User user = userService.getInfoOfMyUser();
+            if (!user.getUsername().equals(username)) {
+                return Response.status(Response.Status.FORBIDDEN)
+                        .entity("{\"message\":\"You can only remove your own likes.\"}")
+                        .build();
+            }
+            boolean removed = reviewService.removeLikeReview(id, ReviewTypes.REVIEW_MEDIA);
+            if (removed)
+                return Response.noContent().build();
+            return Response.status(Response.Status.NOT_FOUND)
+                    .entity("{\"message\":\"List is not liked.\"}").build();
+        } catch (UserNotLoggedException e) {
+            return Response.status(Response.Status.UNAUTHORIZED)
+                    .entity(e.getMessage()).build();
+        }
     }
 
     // Returns like status for a specific review for a user
@@ -271,7 +282,7 @@ public class ReviewController {
                     .build().toString();
             return Response.ok(new UserReviewDto(reviewId, username, uri, uriInfo)).build();
         }
-        return Response.noContent().build();
+        return Response.status(Response.Status.NOT_FOUND).entity("User has not liked the review.").build();
     }
 
     // Return all likes for a review
